+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * `tt coredump inspect`: allows archive path as an argument (archive should be
   created with `tt coredump pack`).
   * `tt coredump inspect`: added `-s` option to specify the location of tarantool sources.
+- `tt cluster publish`: ability to publish a new instance config.
 
 ### Added